style: move listbox to main package level, add style.Theme, and implement style/basic using style.Theme making it easier to create new styles.
