+++ style/style.go
@@ -1,5 +1,11 @@
 package style
 
+import (
+	"github.com/aykevl/tinygl"
+	"github.com/aykevl/tinygl/font"
+	"tinygo.org/x/drivers/pixel"
+)
+
 type Scale uint8
 
 func NewScale(percent int) Scale {
@@ -8,4 +14,15 @@ func NewScale(percent int) Scale {
 
 func (s Scale) Percent() int {
 	return int(s) * 25
+}
+
+type Theme[T pixel.Color] struct {
+	Screen *tinygl.Screen[T]
+
+	// Theme related properties.
+	Font       font.Font
+	Foreground T // text, borders, etc
+	Background T // background
+	Tint       T // highlights (checked boxes, active list element, etc)
+	Scale      Scale
 }
